+++ client/src/common/components/Icons/Back/Back.tsx
@@ -1,20 +1,21 @@
 import React from 'react';
 import {Path} from 'react-native-svg';
+import {IconType} from '..';
 import {COLORS} from '../../../constants/colors';
 import Icon from '../Icon';
 
-export const ForwardIcon = ({fill = COLORS.GREY}) => (
+export const BackIcon: IconType = ({fill = COLORS.GREY}) => (
   <Icon>
     <Path
       fillRule="evenodd"
       clipRule="evenodd"
-      d="M13.357 24.976a.617.617 0 0 1-.425-.764c.676-2.351.917-4.806.708-7.243l-.201-1.566a24.167 24.167 0 0 1-2.603.002c-1.657-.09-3.776-.392-5.325-1.3-2.087-1.218-3.173-2.71-3.444-4.25-.269-1.531.293-2.979 1.237-4.06C4.194 4.768 5.61 3.999 7.194 4c1.613.002 3.304.803 4.716 2.737 1 1.365 1.66 3.33 2.069 4.894a29.703 29.703 0 0 1 .53 2.453 102.275 102.275 0 0 0 3.994-.415c2.65-.33 5.742-.834 7.704-1.534a.617.617 0 1 1 .42 1.163c-2.106.751-5.324 1.268-7.97 1.597a103.18 103.18 0 0 1-3.979.414l.195 1.516a21.245 21.245 0 0 1-.748 7.727.621.621 0 0 1-.768.424Zm-.092-10.8-.055-.301a28.455 28.455 0 0 0-.432-1.933c-.403-1.539-1.017-3.311-1.872-4.478-1.218-1.668-2.559-2.227-3.713-2.229-1.182-.002-2.268.58-2.95 1.368l-.002.002c-.752.86-1.143 1.945-.95 3.037.19 1.085.978 2.306 2.849 3.397h.001c1.279.752 3.144 1.044 4.763 1.132a22.94 22.94 0 0 0 2.36.004Z"
+      d="M17.883 5.055c.369.152.546.573.394.943a19.728 19.728 0 0 0-1.469 6.985l.025 1.45a23.684 23.684 0 0 1 2.448.275c1.611.268 3.67.795 5.105 1.87 1.924 1.433 2.846 3.043 2.946 4.623.099 1.572-.625 2.962-1.683 3.938-1.002.933-2.5 1.548-4.082 1.37-1.614-.18-3.206-1.173-4.39-3.248-.839-1.466-1.273-3.482-1.503-5.063a29.618 29.618 0 0 1-.245-2.37 101.004 101.004 0 0 0-3.836-.021c-2.621.037-5.684.198-7.66.67a.722.722 0 0 1-.336-1.405c2.15-.513 5.358-.673 7.975-.71a101.054 101.054 0 0 1 3.815.02l-.024-1.39v-.03c.064-2.58.599-5.125 1.578-7.513a.722.722 0 0 1 .942-.394ZM16.88 15.883l.204.015c.5.037 1.192.107 1.96.235 1.568.26 3.333.745 4.477 1.603l.002.001c1.689 1.258 2.301 2.514 2.367 3.555.066 1.052-.414 2.041-1.222 2.786l-.002.003c-.73.68-1.82 1.119-2.938.993-1.085-.121-2.304-.788-3.296-2.53-.697-1.218-1.103-3.004-1.328-4.554a28.159 28.159 0 0 1-.224-2.106Z"
       fill={fill}
     />
     <Path
       fillRule="evenodd"
       clipRule="evenodd"
-      d="M24.652 18.215c-.255-.198-.333-.615-.174-.933l2.24-4.467-3.296-2.806c-.248-.211-.311-.633-.142-.941.17-.31.51-.388.757-.177l3.726 3.173c.24.205.309.61.154.917L25.402 18c-.159.318-.495.415-.75.216Z"
+      d="M6.15 11.08a.722.722 0 0 1 .12 1.015l-2.867 3.634 3.307 2.844a.722.722 0 1 1-.942 1.095l-3.831-3.295a.722.722 0 0 1-.097-.995L5.135 11.2a.722.722 0 0 1 1.014-.12Z"
       fill={fill}
     />
   </Icon>
