+++ client/src/common/components/Icons/index.ts
@@ -1,3 +1,10 @@
+import React from 'react';
+import {ColorValue} from 'react-native';
+
+export type IconType = React.FC<{
+  fill?: ColorValue;
+}>;
+
 export * from './Back/Back';
 export * from './Forward/Forward';
 export * from './Home/Home';
